% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bin_width.R
\name{calc_all_bin_width}
\alias{calc_all_bin_width}
\title{Calculate all bin width for a histogram using various recommended methods}
\usage{
calc_all_bin_width(x)
}
\arguments{
\item{x}{A numeric object}
}
\value{
A tibble
}
\description{
Applies `laviz::calc_bin_width` and returns a tibble with all the bin widths
calculated by all the methods.
}
\examples{
library(ggplot2)

calc_all_bin_width(diamonds$carat)

calc_all_bin_width(economics_long$value)

}
\references{
https://en.wikipedia.org/wiki/Histogram#Square-root_choice
https://stats.stackexchange.com/questions/798/calculating-optimal-number-of-bins-in-a-histogram
https://arxiv.org/pdf/1612.07216.pdf
https://cran.r-project.org/web/packages/essHist/essHist.pdf
https://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.170.173&rep=rep1&type=pdf
https://www.neuralengine.org//res/histogram.html
https://www.qimacros.com/histogram-excel/how-to-determine-histogram-bin-interval/#:~:text=Here's%20How%20to%20Calculate%20the,data%20points%20and%20round%20up.
}
