% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-master.R
\name{collection_names}
\alias{collection_names}
\title{List all available palette collections}
\usage{
collection_names()
}
\value{
A character vector of collection names
}
\description{
List all available palette collections
}
\examples{
collection_names()
}
