% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-master.R
\docType{data}
\name{color_palettes}
\alias{color_palettes}
\title{Master color palette list}
\format{
A named list of lists of character vectors (hex color codes)
}
\usage{
color_palettes
}
\description{
Master color palette list
}
\keyword{internal}
