% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drsimonj_palettes.R
\name{drsimonj_pal}
\alias{drsimonj_pal}
\title{Return function to interpolate a drsimonj color palette}
\usage{
drsimonj_pal(palette = "main", alpha = 1, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{alpha}{transparency}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a drsimonj color palette
}
\examples{
drsimonj_pal("cool")

library(scales)
scales::show_col(drsimonj_pal("cool")(10))

filled.contour(volcano, color.palette = drsimonj_pal(), asp = 1)
}
