% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_histo.R
\name{gg_histo}
\alias{gg_histo}
\title{Histograms with options for auto-binning}
\usage{
gg_histo(data, var, binw_select = "FD", facet = NULL, alpha = 0.8,
  fill = "steelblue", colour = "black",
  theme_for_histo = theme_minimal(), title = NULL, subtitle = NULL)
}
\arguments{
\item{data}{Data frame or tibble}

\item{var}{Variable to be plotted}

\item{binw_select}{Specify method to calculate the bin width. "FD" for
Freedman-Diaconis (1981) (default), "Sturges" for Sturges (1926),
"Scott" for Scott (1979), "Square-root" for Square-root (N/A), or
"Rice" for Rice (1944).}

\item{facet}{Variable name to facet wrap by. Default is NULL.}

\item{alpha}{How solid / transparent should the fill be.}

\item{fill}{Fill colour for histrogram.}

\item{colour}{Outline colour of histogram blocks.}

\item{theme_for_histo}{Specify the ggplot theme. Default is theme_minimal().}

\item{title}{The main title of the plot}

\item{subtitle}{A subtitle for the plot}
}
\description{
Function to plot histograms with specified bin width. See reference:
\url{https://www.answerminer.com/blog/binning-guide-ideal-histogram}.
}
\examples{
library(tidyverse)

gg_histo(data = diamonds,
         var = price,
         binw_select = "Sturges",
         facet = NULL)

gg_histo(data = diamonds,
         var = price,
         binw_select = "Scott",
         facet = clarity)

gg_histo(data = diamonds,
         var = price,
         binw_select = "FD",
         facet = clarity)

gg_histo(data = diamonds,
         var = price,
         binw_select = "Square-root",
         facet = clarity)

ggplot(data = diamonds, aes(x = price)) +
  geom_density(alpha = 0.4,
               fill = "darkgreen") +
  theme_minimal()

}
