% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pie.R
\name{gg_pie}
\alias{gg_pie}
\title{A tasteful pie chart}
\usage{
gg_pie(data, var, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{var}{Variable with the data to be sliced}

\item{...}{Additional function arguments}
}
\description{
Sometimes you gotta make a pie chart. And if you do, it should make some
efforts to hellp it to convey information clearly. Some critics say that pie
charts should never be used because the area representations (i.e. pie
slices) are difficuly for humans to make sense of. I agree and I disagree. I
avoid them as much as I can with my own data viz, bar charts seem to work
better to me. But sometimes you have a collaborator who really wants a pie
chart, and you want to be a team player. Also, with less than 4 groups, I
think that they do just fine. Thought that should come witha couple of
caveats: (1) if the pie slices are really small then it becomes really
difficult to tell what value they represent, and (2) if mulitple pie slices
are close in size the it becomes difficult to tell them apart.
}
\examples{
gg_pie(data = ggplot2::mpg,
       var = class)
}
