% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkm.R
\name{ggkm}
\alias{ggkm}
\title{Kaplan-Meier survival plot}
\usage{
ggkm(sfit, main = "Kaplan-Meier Plot", subtitle = NA, risk_table = TRUE,
  timeby = 100, xlims = c(0, max(sfit$time)), ylims = c(0, 1.01),
  xlabs = "Time", ylabs = "Survival Probability (\%)", ystratalabs = NULL,
  ystrataname = NULL, pval = TRUE, pval_threshold = 0.001,
  pvalpos = c(max(sfit$time)/6, 0.2), text_annotate = NA, marks = FALSE,
  shape = 3, linetype = FALSE, linecolor = TRUE, blackandwhite = FALSE,
  palette = c("#0073C2", "#EFC000", "#868686", "#CD534C", "#7AA6DC",
  "#003C67", "#8F7700", "#3B3B3B", "#A73030", "#4A6990"), legend = TRUE,
  legpos = c(0.9, 0.8), subs = NULL, font_family = "",
  adj_table_title = -0.2, adj_y_axis_label = -12.5,
  surv_plot_height = 0.75, risk_table_height = 0.25)
}
\arguments{
\item{sfit}{A survfit object}

\item{main}{String; Main plot title. Default is "Kaplan-Meier Plot".}

\item{subtitle}{String; Subtitle for the plot. Default is NA.}

\item{risk_table}{Logical; \code{TRUE} will show the risk table. Default is
\code{FALSE}}

\item{timeby}{Integer; amount to break the x-axis by.}

\item{xlims}{A numeric vector; The minimum and maximum values for the x-axis.
Default is \code{0} and max time in the survfit obeject.}

\item{ylims}{A numeric vector; The minimum and maximum values for the y-axis.
Default is \code{0} to \code{1.01} to give a little space at the top.}

\item{xlabs}{String; Label for the x-axis. Default is "Time".}

\item{ylabs}{String; Label for the y-axis. Default is "Survival Probability
(\%)"}

\item{ystratalabs}{String vector; Labels for the strata}

\item{ystrataname}{String; The title for the strata.}

\item{pval}{Logical; \code{TRUE} is default and shows a p-value onn the plot.}

\item{pval_threshold}{Numeric; threshold to showing actual p-value or "P <
...". Defualt is \code{0.001}.}

\item{pvalpos}{A numeric vector; Position for the p-value on the plot if
\code{pval} is \code{TRUE}. Default is in the lower left corner of the plot.}

\item{text_annotate}{String; Text to show instead of p-value.}

\item{marks}{Logical; \code{TRUE} will show marks for censoring. Default is
\code{FALSE}}

\item{shape}{Integer; Controls the shape of the censoring mark if \code{TRUE}.
Default is \code{3}, a cross shape. Other good option is \code{4} = X}

\item{linetype}{Logical; if \code{TRUE} then lines will be different types:
solid, dashed, dotted, etc. Default is \code{FALSE}.}

\item{linecolor}{Logical; if \code{TRUE} then lines are different colors.
Default is \code{TRUE}.}

\item{blackandwhite}{Logical; if \code{TRUE}, plot is made in black and white
using R's \code{gray.colors}}

\item{palette}{The color paletter to be used. Default are opinionated
choices.}

\item{legend}{Logical; if \code{TRUE} then the legend is shown.
Default is \code{TRUE}.}

\item{legpos}{Numeric vector; Position of the legend. Default is the upper
right corner of the plot.}

\item{subs}{Default is NULL}

\item{font_family}{String name of the font to be used; Default is "Arial".}

\item{adj_table_title}{Numeric; adjusts the postion of the title for the Risk
Table when used. Combining plots will move this label around. This argument
will artifically move the label out of place so that it ends up in the
right place. Default is \code{-0.20}. __More negative moves left__. This
takes trial and error.}

\item{adj_y_axis_label}{Numeric; adjusts the postion of the title for the
Risk Table when used. Combining plots will move this label around. This
argument will artifically move the label out of place so that it ends up
in the right place. Default is \code{-12.5}. __More negative moves right__.
This takes trial and error.}

\item{surv_plot_height}{the height of the survival plot on the grid. Default
is \code{0.75}. Ignored when risk.table = \code{FALSE}.}

\item{risk_table_height}{the height of the risk table on the grid. Increase
the value when you have many strata. Default is \code{0.25}. Ignored when
risk.table = \code{FALSE}.}
}
\value{
A plot
}
\description{
I mostly user the package \code{survminer} when making quick and easy
survival plots. But when I want to make a plot for publication or for higher
quality, I will use this function \code{ggkm}. Most of the code was cobbled
together from many different versions of this function that people have
shared online. Just google "ggkm" and you'll see them. My version allows me
to make the tweaks and adjustments that I want when making these plots.
}
\examples{
library(survival)
#### Load fonts --------------------------------
# extrafont::fonts()  # Vector of font family names
# extrafont::fonttable()  # Show entire table
# extrafont::font_import()  # imports fonts installed on the system
extrafont::loadfonts(device = "win", quiet = TRUE)
data(colon)
fit <- survfit(Surv(time, status) ~ 1, data = colon)
ggkm(fit, timeby = 500, marks = TRUE)
ggkm(fit, timeby = 500)

fit <- survfit(Surv(time, status) ~ rx, data = colon)
ggkm(fit,
     timeby = 500,
     marks = TRUE,
     ystratalabs = c("Observed",
                     "Level",
                     "Level+"),
     adj_table_title = -0.12)

sfit <- survfit(Surv(time, status) ~ rx, data = colon)
foo <- ggkm(sfit = sfit,
            risk_table = TRUE,
            xlabs = "Time (in months)",
            ylabs = "Survival Probability (\%)",
            xlims = c(0, max(sfit$time)),
            ylims = c(0, 1.01),
            ystratalabs = NULL,
            ystrataname = NULL,
            timeby = 500,
            main = "Kaplan-Meier Plot",
            pval = FALSE,
            marks = TRUE,
            linetype = FALSE,
            linecolor = TRUE,
            shape = 3,  # 3 = cross, 4 = X
            legend = TRUE,
            subs = NULL,
            legpos = c(1, 1),
            text_annotate = "Here is some text",
            pvalpos = c(500, 0.25),
            font_family = "Times New Roman")

foo

}
