% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_boxplot_legend.R
\name{ggplot_box_legend}
\alias{ggplot_box_legend}
\title{Legend for a box plot}
\usage{
ggplot_box_legend(title = "Boxplot legend", family = "")
}
\arguments{
\item{title}{The text for the title}

\item{family}{Font family; may need to load `extrafont` package first}
}
\value{
A ggplot object
}
\description{
A figure legend that explains the components and symbols for a box plot.
}
\examples{
library(ggplot2)
ggplot_box_legend()

}
\references{
https://owi.usgs.gov/blog/boxplots/
}
