% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogg.R
\name{histogg}
\alias{histogg}
\title{histogg}
\usage{
histogg(data, var, binw_select = "FD", facet = NULL, alpha = 0.8,
  fill = "steelblue", colour = "black", theme_for_histo = theme_minimal())
}
\arguments{
\item{data}{Data frame or tibble}

\item{var}{Variable to be plotted}

\item{binw_select}{Specify method to calculate the bin width. "FD" for
Freedman-Diaconis (1981) (default), "Sturges" for Sturges (1926),
"Scott" for Scott (1979), "Square-root" for Square-root (N/A), or
"Rice" for Rice (1944).}

\item{facet}{Variable name to facet wrap by. Default is NULL.}

\item{alpha}{How solid / transparent should the fill be.}

\item{fill}{Fill colour for histrogram.}

\item{colour}{Outline colour of histogram blocks.}

\item{theme_for_histo}{Specify the ggplot theme. Default is theme_minimal().}
}
\description{
Function to plot histograms with specified bin width. See reference:
\url{https://www.answerminer.com/blog/binning-guide-ideal-histogram}.
}
\examples{
pacman::p_load(tidyverse)

histogg(data = diamonds,
        var = price,
        binw_select = "FD",
        facet = NULL)

histogg(data = diamonds,
        var = price,
        binw_select = "Sturges",
        facet = NULL)

histogg(data = diamonds,
        var = price,
        binw_select = "Scott",
        facet = NULL)

histogg(data = diamonds,
        var = price,
        binw_select = "Square-root",
        facet = NULL)

histogg(data = diamonds,
        var = price,
        binw_select = "Rice",
        facet = NULL)

ggplot(data = diamonds, aes(x = price)) +
  geom_density(alpha = 0.4,
               fill = "darkgreen") +
  theme_minimal()

}
