% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laviz.R
\docType{package}
\name{laviz-package}
\alias{laviz}
\alias{laviz-package}
\title{laviz: an opinionated package of Emile's visualizations}
\description{
The objective of this package is to perform statistical inference using a
grammar that illustrates the underlying concepts and a format that coheres
with the tidyverse.
}
\examples{
# Example usage:
library(laviz)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emilelatour/laviz}
  \item Report bugs at \url{https://github.com/emilelatour/laviz/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
