% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laviz.R
\docType{package}
\name{laviz}
\alias{laviz}
\title{laviz: an opinionated package of Emile's visualizations}
\description{
The objective of this package is to perform statistical inference
using a grammar that illustrates the underlying
concepts and a format that coheres with the tidyverse.
}
\examples{
# Example usage:
library(laviz)
}
