% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-palettes.R
\name{list_palettes}
\alias{list_palettes}
\title{List available palettes}
\usage{
list_palettes(n = 10)
}
\arguments{
\item{n}{Number of rows to print (default: 10). Use `Inf` to print all rows.}
}
\value{
A tibble with collection, palette, and n
}
\description{
List available palettes
}
\examples{
list_palettes(20)
list_palettes(Inf)
}
