% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsu_palettes.R
\name{ohsu_cols}
\alias{ohsu_cols}
\title{OHSU colors}
\usage{
ohsu_cols(...)
}
\arguments{
\item{...}{Character names of ohsu_colors}
}
\description{
Function to extract ohsu colors as hex codes

\tabular{ll}{
  horizon yellow    \tab #FFC939\cr
  marquam blue      \tab #5E97C9\cr
  ohsu gray         \tab #585E60\cr
  terwilliger green \tab #57B147\cr
  dark blue         \tab #002776\cr
  orange            \tab #ED6D23\cr
  pink red          \tab #ED1941\cr
  cyan blue         \tab #0E4D8F\cr
  red               \tab #C34D36
}
}
\examples{
library(ggplot2)
ohsu_cols()
ohsu_cols("horizon yellow")
ohsu_cols("horizon yellow", "terwilliger green")
ohsu_cols("terwilliger green", "horizon yellow")

ggplot(data = mtcars,
       aes(x = hp,
           y = mpg)) +
  geom_point(color = ohsu_cols("terwilliger green"),
             size = 4,
             alpha = 0.8) +
  theme_minimal()
}
