% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsu_palettes.R
\name{ohsu_pal}
\alias{ohsu_pal}
\title{Return function to interpolate an OHSU color palette}
\usage{
ohsu_pal(palette = "main", alpha = 1, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ohsu_palettes}

\item{alpha}{transparency}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an OHSU color palette
}
\examples{
ohsu_pal("cool")

library(scales)
scales::show_col(ohsu_pal("cool")(10))
scales::show_col(ohsu_pal("main")(4))
scales::show_col(ohsu_pal("full")(9))
scales::show_col(ohsu_pal("cool")(4))
scales::show_col(ohsu_pal("hot")(4))
scales::show_col(ohsu_pal("mixed")(5))
scales::show_col(ohsu_pal("grey")(4))

filled.contour(volcano, color.palette = ohsu_pal(), asp = 1)
}
