% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-ohsu.R
\docType{data}
\name{ohsu_palettes}
\alias{ohsu_palettes}
\title{OHSU corporate color palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
ohsu_palettes
}
\description{
OHSU brand colors and named palettes for visualization.
}
\keyword{datasets}
