% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-master.R
\name{palette_names}
\alias{palette_names}
\title{List available palette names from a collection}
\usage{
palette_names(collection)
}
\arguments{
\item{collection}{Name of the collection (e.g., "sb", "ohsu")}
}
\value{
A character vector of palette names
}
\description{
List available palette names from a collection
}
\examples{
palette_names("sb")
# palette_names("personal")
}
