% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-personal.R
\docType{data}
\name{personal_palettes}
\alias{personal_palettes}
\title{Personal color palettes}
\format{
A named list of character vectors (hex color codes).
}
\usage{
personal_palettes
}
\description{
A collection of custom color palettes Emile has used or adapted for personal and professional work.
}
\keyword{internal}
