% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-sb.R
\docType{data}
\name{sb_palettes}
\alias{sb_palettes}
\title{SB Color Palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
sb_palettes
}
\description{
Palettes adapted from seaborn colorblind-friendly sets.
}
\keyword{datasets}
