% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_color_collection}
\alias{scale_color_collection}
\alias{scale_fill_collection}
\title{Generic color scale from any palette collection}
\usage{
scale_color_collection(collection, palette, discrete = TRUE, ...)

scale_fill_collection(collection, palette, discrete = TRUE, ...)
}
\arguments{
\item{collection}{Name of palette collection (e.g. "sb", "ohsu")}

\item{palette}{Name of palette in that collection}

\item{discrete}{Logical: use discrete scale?}

\item{...}{Other arguments passed to `scale_*_manual` or `scale_*_gradientn`}
}
\description{
Generic color scale from any palette collection
}
