% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drsimonj_palettes.R
\name{scale_color_drsimonj}
\alias{scale_color_drsimonj}
\alias{scale_colour_drsimonj}
\title{Color scale constructor for drsimonj colors}
\usage{
scale_color_drsimonj(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)

scale_colour_drsimonj(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for drsimonj colors
}
\examples{
library(ggplot2)
# Color by discrete variable using default palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
    geom_point(size = 4) +
    scale_color_drsimonj()

# Color by numeric variable with cool palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
    geom_point(size = 4, alpha = .6) +
    scale_color_drsimonj(discrete = FALSE, palette = "cool")
}
