% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{scale_color_sb}
\alias{scale_color_sb}
\alias{scale_fill_sb}
\title{Shortcut for SB ggplot2 color and fill scales}
\usage{
scale_color_sb(palette, discrete = TRUE, ...)

scale_fill_sb(palette, discrete = TRUE, ...)
}
\arguments{
\item{palette}{Name of palette in that collection}

\item{discrete}{Logical: use discrete scale?}

\item{...}{Other arguments passed to `scale_*_manual` or `scale_*_gradientn`}
}
\description{
Uses `scale_color_collection()` and `scale_fill_collection()` with `"sb"` as the collection.
}
