% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_comma.R
\name{scale_comma}
\alias{scale_comma}
\alias{scale_x_comma}
\alias{scale_y_comma}
\title{Axis formatting with commas}
\usage{
scale_x_comma(
  ...,
  labels = scales::comma_format(accuracy = 1, scale = 1, prefix = "", suffix = "",
    big.mark = ",", decimal.mark = ".", trim = TRUE)
)

scale_y_comma(
  ...,
  labels = scales::comma_format(accuracy = 1, scale = 1, prefix = "", suffix = "",
    big.mark = ",", decimal.mark = ".", trim = TRUE)
)
}
\arguments{
\item{...}{Arguments to `ggplot2::scale_x_continuous` and `ggplot2::scale_x_continuous`}

\item{labels}{default is `scales::percent` See `?ggplot2::scale_x_continuous` for additional details}
}
\description{
Wrappers around `ggplot2::scale_x_continuous` and
`ggplot2::scale_y_continuous` to make it easy to format with commas
}
\examples{
library(dplyr)
library(ggplot2)
library(palmerpenguins)

penguins \%>\%
  ggplot(data = .,
         aes(x = flipper_length_mm,
             y = body_mass_g,
             colour = species,
             shape = species)) +
  geom_point()

penguins \%>\%
  ggplot(data = .,
         aes(x = flipper_length_mm,
             y = body_mass_g,
             colour = species,
             shape = species)) +
  geom_point() +
  scale_y_comma()

}
\references{
https://github.com/thomas-neitmann/scalesextra
}
