% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wom_palettes.R
\name{scale_fill_wom}
\alias{scale_fill_wom}
\title{Fill scale constructor for wom colors}
\usage{
scale_fill_wom(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in wom_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for wom colors
}
\examples{
library(ggplot2)
# Fill by discrete variable with different palette + remove legend (guide)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
    geom_bar() +
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    scale_fill_wom(palette = "mixed", guide = "none")
}
