% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_pct.R
\name{scale_continuous}
\alias{scale_continuous}
\alias{scale_x_pct}
\alias{scale_y_pct}
\title{Axis scales and formatting for percentages}
\usage{
scale_x_pct(..., limits = c(0, 1), labels = scales::percent)

scale_y_pct(..., limits = c(0, 1), labels = scales::percent)
}
\arguments{
\item{...}{Arguments to `ggplot2::scale_x_continuous` and `ggplot2::scale_x_continuous`}

\item{limits}{default is `c(0, 1)` See `?ggplot2::scale_x_continuous` for additional details}

\item{labels}{default is `scales::percent` See `?ggplot2::scale_x_continuous` for additional details}
}
\description{
Wrappers around `ggplot2::scale_x_continuous` and
`ggplot2::scale_y_continuous` to make it easy to format percentages.
}
\examples{
library(dplyr)
library(ggplot2)
library(palmerpenguins)

penguins \%>\%
  dplyr::count(species) \%>\%
  mutate(pct = n / sum(n)) \%>\%
  ggplot(data = .,
         aes(x = species,
             y = pct)) +
  geom_bar(stat = "identity")

penguins \%>\%
  dplyr::count(species) \%>\%
  mutate(pct = n / sum(n)) \%>\%
  ggplot(data = .,
         aes(x = species,
             y = pct)) +
  geom_bar(stat = "identity") +
  scale_y_pct()

}
\references{
https://github.com/thomas-neitmann/scalesextra
}
