% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_tight.R
\name{scale_tight}
\alias{scale_tight}
\alias{scale_x_tight}
\alias{scale_y_tight}
\title{Axis formatting tight to the plot}
\usage{
scale_x_tight(...)

scale_y_tight(...)
}
\arguments{
\item{...}{Arguments to `ggplot2::scale_x_continuous` and `ggplot2::scale_x_continuous`}
}
\description{
Wrappers around `ggplot2::scale_x_continuous` and
`ggplot2::scale_y_continuous` to make it easy to tighten up the axis
}
\examples{
library(dplyr)
library(ggplot2)
library(palmerpenguins)

penguins \%>\%
  ggplot(data = .,
         aes(x = flipper_length_mm,
             y = body_mass_g,
             colour = species,
             shape = species)) +
  geom_point()

penguins \%>\%
  ggplot(data = .,
         aes(x = flipper_length_mm,
             y = body_mass_g,
             colour = species,
             shape = species)) +
  geom_point() +
  scale_x_tight()


penguins \%>\%
  dplyr::count(species) \%>\%
  mutate(pct = n / sum(n)) \%>\%
  ggplot(data = .,
         aes(x = species,
             y = pct)) +
  geom_bar(stat = "identity")

penguins \%>\%
  dplyr::count(species) \%>\%
  mutate(pct = n / sum(n)) \%>\%
  ggplot(data = .,
         aes(x = species,
             y = pct)) +
  geom_bar(stat = "identity") +
  scale_y_tight()

}
\references{
https://github.com/thomas-neitmann/scalesextra
https://stackoverflow.com/questions/44170871/how-does-ggplot-scale-continuous-expand-argument-work
}
