% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{show_palette}
\alias{show_palette}
\title{Show a palette as a color bar}
\usage{
show_palette(palette, border = "grey80")
}
\arguments{
\item{palette}{A character vector of hex color codes.}

\item{border}{Optional border color for color boxes.}
}
\description{
Displays a horizontal strip of colors using `scales::show_col()`, useful for
continuous palettes or clean comparisons.
}
\examples{
show_palette(get_palette("sb", "deep", type = "continuous", n = 20))
}
