% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette}
\alias{view_palette}
\title{View a palette with color labels}
\usage{
view_palette(palette)
}
\arguments{
\item{palette}{A character vector of hex color codes.}
}
\description{
Displays a pie chart with labeled hex codes for visualizing and identifying
individual colors in a palette.
}
\examples{
view_palette(get_palette("sb", "deep"))
}
