% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{view_palette_collection}
\alias{view_palette_collection}
\title{View all palettes in a collection}
\usage{
view_palette_collection(
  collection,
  type = "discrete",
  n = 20,
  border = "grey80"
)
}
\arguments{
\item{collection}{Name of the collection (e.g., "sb", "ohsu")}

\item{type}{One of "discrete" or "continuous"}

\item{n}{If type = "continuous", number of interpolated colors to show}

\item{border}{Border color (passed to `scales::show_col()`)}
}
\description{
Visualizes each palette in a collection using `scales::show_col()`.
}
\examples{
view_palette_collection("sb")                             # discrete
view_palette_collection("sb", type = "continuous", n = 40)  # continuous
}
