% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-shape-and-linetypes.R
\name{view_shapes}
\alias{view_shapes}
\alias{view_linetypes}
\title{Dislpay the shapes and linetypes with codes}
\usage{
view_shapes()

view_linetypes()
}
\value{
A plot
}
\description{
Quick way to display the shapes and linetypes available in base R and ggplot2
with their corresponding codes
}
\examples{
\dontrun{
view_shapes()
view_linetypes()
}
}
