% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wom_palettes.R
\name{wom_cols}
\alias{wom_cols}
\title{WoM colors}
\usage{
wom_cols(...)
}
\arguments{
\item{...}{Character names of wom_cols}
}
\description{
Function to extract wom colors as hex codes

\tabular{ll}{
  blue      \tab #007bff\cr
  indigo    \tab #6610f2\cr
  purple    \tab #6f42c1\cr
  pink      \tab #e83e8c\cr
  red       \tab #dc3545\cr
  orange    \tab #fd7e14\cr
  yellow    \tab #ffc107\cr
  green     \tab #28a745\cr
  teal      \tab #20c997\cr
  cyan      \tab #17a2b8\cr
  white     \tab #ffffff\cr
  gray      \tab #6c757d\cr
  graydark  \tab #343a40\cr
  primary   \tab #007bff\cr
  secondary \tab #6c757d\cr
  success   \tab #28a745\cr
  info      \tab #17a2b8\cr
  warning   \tab #ffc107\cr
  danger    \tab #dc3545\cr
  light     \tab #f8f9fa\cr
  dark      \tab #343a40
}
}
\examples{
library(ggplot2)
wom_cols()
wom_cols("red")
wom_cols("red", "blue")
wom_cols("blue", "red")

ggplot(data = mtcars,
       aes(x = hp,
           y = mpg)) +
  geom_point(color = wom_cols("red"),
             size = 4,
             alpha = 0.8) +
  theme_minimal()
}
