% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wom_palettes.R
\name{wom_pal}
\alias{wom_pal}
\title{Return function to interpolate a wom color palette}
\usage{
wom_pal(palette = "main", alpha = 1, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in wom_palettes}

\item{alpha}{transparency}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a wom color palette
}
\examples{
wom_pal("cool")

library(scales)
scales::show_col(wom_pal("cool")(10))

filled.contour(volcano, color.palette = wom_pal(), asp = 1)
}
