% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drsimonj_palettes.R
\name{drsimonj_cols}
\alias{drsimonj_cols}
\title{Function to extract drsimonj colors as hex codes}
\usage{
drsimonj_cols(...)
}
\arguments{
\item{...}{Character names of drsimonj_colors}
}
\description{
Function to extract drsimonj colors as hex codes
}
\examples{
library(ggplot2)
drsimonj_cols()
drsimonj_cols("red")
drsimonj_cols("red", "blue")
drsimonj_cols("blue", "red")

ggplot(data = mtcars,
       aes(x = hp,
           y = mpg)) +
  geom_point(color = drsimonj_cols("red"),
             size = 4,
             alpha = 0.8) +
  theme_minimal()
}
