% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-drsimonj.R
\docType{data}
\name{drsimonj_palettes}
\alias{drsimonj_palettes}
\title{drsimonj color palettes}
\format{
A named list of character vectors (hex color codes).
}
\usage{
drsimonj_palettes
}
\description{
A compact set of color palettes inspired by [drsimonj’s blog post](https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2).
}
\keyword{internal}
