% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohsu_palettes.R
\name{ohsu_cols}
\alias{ohsu_cols}
\title{Function to extract ohsu colors as hex codes}
\usage{
ohsu_cols(...)
}
\arguments{
\item{...}{Character names of ohsu_colors}
}
\description{
Function to extract ohsu colors as hex codes
}
\examples{
library(ggplot2)
ohsu_cols()
ohsu_cols("horizon yellow")
ohsu_cols("horizon yellow", "terwilliger green")
ohsu_cols("terwilliger green", "horizon yellow")

ggplot(data = mtcars,
       aes(x = hp,
           y = mpg)) +
  geom_point(color = ohsu_cols("terwilliger green"),
             size = 4,
             alpha = 0.8) +
  theme_minimal()
}
