% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-wom.R
\docType{data}
\name{wom_palettes}
\alias{wom_palettes}
\title{WoM color palettes}
\format{
A named list of character vectors (hex color codes).
}
\usage{
wom_palettes
}
\description{
Bootstrap-based color palettes used in WoM-themed graphics.
}
\details{
These palettes are based on Bootstrap's standard and semantic color naming.
}
\keyword{internal}
